requirements.txt
markdown-it-py==2.2.0
mdurl==0.1.2
Pygments==2.15.1
rich==13.3.5
aiogram==3.0.0b7
setuptools~=57.0.0
aiokafka~=0.8.0
pydantic~=1.10.7
PyYAML~=6.0
critical~=0.0.1
typer~=0.9.0
redis~=4.5.5